Changed the variable 'twentyBanknotes' from int type to double

+++ beginner/uri_1003/src/bee_1018/Banknotes.java
@@ -33,7 +33,7 @@ public class Banknotes {
         int fiftyBanknotes = restOfOneHundredBanknotes / 50 ;
 
 
-        int restOfFiftyBanknotes = ((restOfOneHundredBanknotes % 50) - 50);
+        double restOfFiftyBanknotes = ((restOfOneHundredBanknotes % 50) - 50);
 
         double twentyBanknotes = restOfFiftyBanknotes  / 20;
         System.out.println("0 nota (s) de R$ 20,00");
